integration tests refactorr
