+++ test/examples/azureOpenaiChatCompletion.js
@@ -1,17 +1,18 @@
-import { setupMonocle } from '../../dist';
-setupMonocle("azure-openai.app");
+const { setupMonocle } = require("../../dist");
 
-import { OpenAI } from "openai";
+setupMonocle("openai.app");
 
+const { OpenAI } = require("openai");
+
+// Create Azure OpenAI client
 const client = new OpenAI({
   apiKey: process.env["AZURE_OPENAI_API_KEY"],
   baseURL: process.env["AZURE_OPENAI_ENDPOINT_1"], // Azure endpoint URL
   defaultQuery: { "api-version": "2023-05-15" }, // API version
-//   defaultHeaders: { "api-key": process.env["AZURE_OPENAI_API_KEY"] }
+  //   defaultHeaders: { "api-key": process.env["AZURE_OPENAI_API_KEY"] }
 });
 
 async function main() {
-
   const chatCompletion = await client.chat.completions.create({
     messages: [
       { role: "user", content: "What is an americano?" },
@@ -66,19 +67,10 @@ async function main() {
   //   console.log("First 5 dimensions:", embedding.slice(0, 5));
 }
 
-if (require.main === module) {
-  (async () => {
-    try {
-      await main();
-    } catch (e) {
-      console.error("Error during processing:", e);
-    }
-    // Wait 5 seconds then exit
-    setTimeout(() => {
-      console.log("Exiting after 5 seconds...");
-      process.exit(0); // force clean exit
-    }, 5_000);
-  })();
-}
+exports.main = main;
 
-export { main };
+if (require.main === module) {
+  main().catch((error) => {
+    console.error("Error occurred:", error);
+  });
+}
