+++ test/examples/bedrockRuntimeClient.ndjson
@@ -1,2 +1,2 @@
-{"name":"aws_bedrock_invoke_model","context":{"trace_id":"e9f6c81556ef888cbd960c0518c867e7","span_id":"e8ed229b71184102","trace_flags":1},"kind":0,"parent_id":"1a3e535df937ca0c","start_time":"2025-03-30T20:26:55.561000Z","end_time":"2025-03-30T20:26:58.823000Z","status":{"status_code":"UNSET"},"attributes":{"monocle_apptrace.version":"0.1.0","monocle_apptrace.language": "js","workflow.name": "bedrock.app","span.type":"inference","entity.1.name":"anthropic.claude-v2","entity.1.type":"model.llm.anthropic.claude-v2","entity.count":1},"events":[{"name":"data.input","timestamp":"2025-03-30T20:26:58.823000Z","attributes":{"input":["\n\nHuman: What is coffee?\n\nAssistant: "]}},{"name":"data.output","timestamp":"2025-03-30T20:26:58.823000Z","attributes":{"response":["Coffee is a brewed drink prepared from roasted coffee beans, the seeds of berries from certain Coffea species. When coffee berries turn from green to bright red in color – indicating ripeness – they are picked, processed, and dried. Dried coffee beans are roasted to varying degrees, depending on the desired flavor, before being ground and brewed to create coffee."]}}],"links":[],"resource":{"attributes":{"service.name":"bedrock.app"}}}
-{"name":"workflow","context":{"trace_id":"e9f6c81556ef888cbd960c0518c867e7","span_id":"1a3e535df937ca0c","trace_flags":1},"kind":0,"parent_id":"None","start_time":"2025-03-30T20:26:55.561000Z","end_time":"2025-03-30T20:26:58.824000Z","status":{"status_code":"UNSET"},"attributes":{"monocle_apptrace.version":"0.1.0","monocle_apptrace.language": "js","workflow.name": "bedrock.app","span.type":"workflow","entity.1.name":"bedrock.app","entity.1.type":"workflow.generic","entity.2.type":"app_hosting.generic","entity.2.name":"generic"},"events":[],"links":[],"resource":{"attributes":{"service.name":"bedrock.app"}}}
+{"name":"aws_bedrock_invoke_model","context":{"trace_id":"e9f6c81556ef888cbd960c0518c867e7","span_id":"e8ed229b71184102","trace_flags":1},"kind":0,"parent_id":"1a3e535df937ca0c","start_time":"2025-03-30T20:26:55.561000Z","end_time":"2025-03-30T20:26:58.823000Z","status":{"status_code":"OK","description":"OK"},"attributes":{"monocle_apptrace.version":"0.1.0","monocle_apptrace.language": "js","workflow.name": "bedrock.app","span.type":"inference","entity.1.name":"anthropic.claude-v2","entity.1.type":"model.llm.anthropic.claude-v2","entity.count":1,"span.source":"default-source-path"},"events":[{"name":"data.input","timestamp":"2025-03-30T20:26:58.823000Z","attributes":{"input":["\n\nHuman: What is coffee?\n\nAssistant: "]}},{"name":"data.output","timestamp":"2025-03-30T20:26:58.823000Z","attributes":{"response":["Coffee is a brewed drink prepared from roasted coffee beans, the seeds of berries from certain Coffea species. When coffee berries turn from green to bright red in color – indicating ripeness – they are picked, processed, and dried. Dried coffee beans are roasted to varying degrees, depending on the desired flavor, before being ground and brewed to create coffee."]}}],"links":[],"resource":{"attributes":{"service.name":"bedrock.app"}}}
+{"name":"workflow","context":{"trace_id":"e9f6c81556ef888cbd960c0518c867e7","span_id":"1a3e535df937ca0c","trace_flags":1},"kind":0,"parent_id":"None","start_time":"2025-03-30T20:26:55.561000Z","end_time":"2025-03-30T20:26:58.824000Z","status":{"status_code":"OK","description":""},"attributes":{"monocle_apptrace.version":"0.1.0","monocle_apptrace.language": "js","workflow.name": "bedrock.app","span.type":"workflow","entity.1.name":"bedrock.app","entity.1.type":"workflow.generic","entity.2.type":"app_hosting.generic","entity.2.name":"generic","span.source":"default-source-path"},"events":[],"links":[],"resource":{"attributes":{"service.name":"bedrock.app"}}}
