+++ test/examples/sagemakerInvokeEndpoint.ndjson
@@ -1,2 +1,4 @@
-{"name":"aws_sagemaker_invoke_endpoint","context":{"trace_id":"fbd3169bdf3cb0cbd792b5e61ac17624","span_id":"ffc626657e952d37","trace_flags":1},"kind":0,"parent_id":"f76fe5784110c8d6","start_time":"2025-03-30T20:22:28.895000Z","end_time":"2025-03-30T20:22:30.351000Z","status":{"status_code":"UNSET"},"attributes":{"monocle_apptrace.version":"0.1.0","monocle_apptrace.language": "js","workflow.name":"sagemaker.app","span.type":"inference","entity.1.name":"okahu-sagemaker-rag-qa-ep","entity.1.type":"model.llm.okahu-sagemaker-rag-qa-ep","entity.count":1},"events":[{"name":"data.input","timestamp":"2025-03-30T20:22:30.351000Z","attributes":{"input":["{\"question\":\"What is coffee?\",\"context\":\"Coffee is a beverage brewed from roasted, ground coffee beans.\"}"]}},{"name":"data.output","timestamp":"2025-03-30T20:22:30.351000Z","attributes":{"response":["a beverage brewed from roasted, ground coffee beans"]}}],"links":[],"resource":{"attributes":{"service.name":"sagemaker.app"}}}
-{"name":"workflow","context":{"trace_id":"fbd3169bdf3cb0cbd792b5e61ac17624","span_id":"f76fe5784110c8d6","trace_flags":1},"kind":0,"parent_id":"None","start_time":"2025-03-30T20:22:28.895000Z","end_time":"2025-03-30T20:22:30.352000Z","status":{"status_code":"UNSET"},"attributes":{"monocle_apptrace.version":"0.1.0","monocle_apptrace.language": "js","workflow.name":"sagemaker.app","span.type":"workflow","entity.1.name":"sagemaker.app","entity.1.type":"workflow.generic","entity.2.type":"app_hosting.generic","entity.2.name":"generic"},"events":[],"links":[],"resource":{"attributes":{"service.name":"sagemaker.app"}}}
+{"name":"aws_sagemaker_invoke_endpoint","context":{"trace_id":"0408bd4ebe168939147a995c854baa5e","span_id":"05c0ff294c656d99","trace_flags":1},"kind":0,"parent_id":"85fa3d3e26937ca8","start_time":"2025-06-25T12:15:19.748000Z","end_time":"2025-06-25T12:15:22.500000Z","status":{"status_code":"OK","description":"OK"},"attributes":{"monocle_apptrace.version":"0.1.2","monocle_apptrace.language":"js","span.source":"default-source-path","workflow.name":"openai.app","span.type":"inference","entity.1.name":"okahu-sagemaker-rag-qa-ep","entity.1.type":"model.llm.okahu-sagemaker-rag-qa-ep","entity.count":1},"events":[{"name":"data.input","timestamp":"2025-06-25T12:15:22.500000Z","attributes":{"input":["{\"question\":\"What is coffee?\",\"context\":\"Coffee is a beverage brewed from roasted, ground coffee beans.\"}"]}},{"name":"data.output","timestamp":"2025-06-25T12:15:22.500000Z","attributes":{"response":["a beverage brewed from roasted, ground coffee beans"],"status":200,"status_code":"success"}}],"links":[],"resource":{"attributes":{"service.name":"openai.app"}}}
+{"name":"workflow","context":{"trace_id":"0408bd4ebe168939147a995c854baa5e","span_id":"85fa3d3e26937ca8","trace_flags":1},"kind":0,"parent_id":"None","start_time":"2025-06-25T12:15:19.747000Z","end_time":"2025-06-25T12:15:22.500000Z","status":{"status_code":"OK","description":""},"attributes":{"span.type":"workflow","entity.1.name":"openai.app","entity.1.type":"workflow.generic","entity.2.type":"app_hosting.generic","entity.2.name":"generic","monocle_apptrace.version":"0.1.2","monocle_apptrace.language":"js","span.source":"default-source-path","workflow.name":"openai.app"},"events":[],"links":[],"resource":{"attributes":{"service.name":"openai.app"}}}
+{"name":"aws_sagemaker_invoke_endpoint","context":{"trace_id":"24310af6ccd454792555ab9b5476c97b","span_id":"d1fe19418261620f","trace_flags":1},"kind":0,"parent_id":"8b75bb85d4375a26","start_time":"2025-06-25T12:15:22.501000Z","end_time":"2025-06-25T12:15:24.445000Z","status":{"status_code":"ERROR","description":"The security token included in the request is invalid."},"attributes":{"monocle_apptrace.version":"0.1.2","monocle_apptrace.language":"js","span.source":"default-source-path","workflow.name":"openai.app","span.type":"inference","entity.1.name":"okahu-sagemaker-rag-qa-ep","entity.1.type":"model.llm.okahu-sagemaker-rag-qa-ep","entity.count":1},"events":[{"name":"data.input","timestamp":"2025-06-25T12:15:24.445000Z","attributes":{"input":["{\"question\":\"What is coffee?\",\"context\":\"Coffee is a beverage brewed from roasted, ground coffee beans.\"}"]}},{"name":"data.output","timestamp":"2025-06-25T12:15:24.445000Z","attributes":{"response":"UnrecognizedClientException: The security token included in the request is invalid.","status":403,"status_code":"error"}}],"links":[],"resource":{"attributes":{"service.name":"openai.app"}}}
+{"name":"workflow","context":{"trace_id":"24310af6ccd454792555ab9b5476c97b","span_id":"8b75bb85d4375a26","trace_flags":1},"kind":0,"parent_id":"None","start_time":"2025-06-25T12:15:22.501000Z","end_time":"2025-06-25T12:15:24.445000Z","status":{"status_code":"OK","description":""},"attributes":{"span.type":"workflow","entity.1.name":"openai.app","entity.1.type":"workflow.generic","entity.2.type":"app_hosting.generic","entity.2.name":"generic","monocle_apptrace.version":"0.1.2","monocle_apptrace.language":"js","span.source":"default-source-path","workflow.name":"openai.app"},"events":[],"links":[],"resource":{"attributes":{"service.name":"openai.app"}}}
