+++ test/integration/azure-openai.test.ts
@@ -36,7 +36,7 @@ describe('Azure OpenAI Sample', () => {
         await (await import('../examples/azureOpenaiChatCompletion')).main();
 
         // Wait for async operations to complete
-        await new Promise(resolve => setTimeout(resolve, 1000));
+        await new Promise(resolve => setTimeout(resolve, 5000));
 
         const sortedExpectedSpans = sortSpans(expectedSpans);
         const sortedCapturedLogs = sortSpans(capturedLogs);
