+++ test/integration/openai.test.ts
@@ -37,7 +37,7 @@ describe('OpenAI Sample', () => {
         await (await import('../examples/openaiChatCompletion')).main()
 
         // Wait for async operations to complete
-        await new Promise(resolve => setTimeout(resolve, 1000));
+        await new Promise(resolve => setTimeout(resolve, 5000));
 
         const sortedExpectedSpans = sortSpans(expectedSpans);
         const sortedCapturedLogs = sortSpans(capturedLogs);
